pyproject.toml
[project]
name = "noter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "passlib[argon2]>=1.7.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[async]>=2.0.45",
]

[dependency-groups]
dev = ["alembic>=1.17.2"]
